contrib/tester-progs: fork-tester
Add a fork-tester program to illustrate issues for processes that fork()
but do not exec().

For these cases, we do not have process information because this is
created in the exec event.

So far, we seem to handle these cases by iterating the process tree and
finding the parent of the process (see event_find_curr()). After a
fork(), the parent has the same properties (other than the pid) as the
parent so the above approach more or less works.

Unless the parent dies. In this case, the process is inherited by pid 1,
which means that the process information we get is the init process
(e.g., systemd).

The fork-tester program implements the above scenario. First it spawns a
child (chlid1) which also spawns a child (child2). child1 then exits,
and child2 is passwed to init 1. Then child2 does a connect.

fork-tester output:
$ ./contrib/tester-progs/fork-tester
parent: (pid:146611, ppid:102949) starts
child 1 (pid:146612) exits
parent: (pid:146611) child (146612) exited with: 0
child 2 (pid:146613, ppid:1) connecting to 8.8.8.8:53
child 2 done

generated event:
{
  "process_connect": {
    "process": {
      "exec_id": "OjExMDAwMDAwMDox",
      "pid": 1,
      "uid": 0,
      "cwd": "/",
      "binary": "/usr/lib/systemd/systemd",
      "arguments": "splash",
      "flags": "procFS auid rootcwd",
      "start_time": "2022-05-03T06:41:48.244Z",
      "auid": 0,
      "refcnt": 138
    },
    "parent": {},
    "source_ip": "192.168.0.66",
    "source_port": 43546,
    "destination_ip": "8.8.8.8",
    "destination_port": 53,
    "protocol": "TCP"
  },
  "time": "2022-05-05T10:48:53.581Z"
}

This commit just illustrates the issue. Upcoming patches will attempt
to fix it.

[commit: 879e179b224e15f25ac3b7c692d81d2c2a2dfab1]

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>