TestFork: initial code for the fork test
Introduce a test that tests that tetragon properly handles events that
come from processes that fork() but do not exec().

This patch just introduces the code that spawns the fork-tester program,
and parses its output. It does not start the agent to test that events
are properly created. See subsequent patch for that.

[commit: 072a24de7225fdb80080222e5e5ffb1132d5c457]

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>